test: 컨트롤러 테스트 분리
